requirement.txt
streamlit>=1.35.0
torch>=2.0.1
torchvision>=0.15.2
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
